Renamed data view(er) to strand view(er).
