--- src/main/java/nl/tudelft/lifetiles/data/controller/package-info.java
@@ -1,6 +1,0 @@
-/**
- * @author Joren Hammudoglu
- *
- */
-package nl.tudelft.lifetiles.data.controller;
-
