--- src/main/java/nl/tudelft/lifetiles/data/model/package-info.java
@@ -1,6 +1,0 @@
-/**
- * @author Joren Hammudoglu
- *
- */
-package nl.tudelft.lifetiles.data.model;
-
