+++ src/main/resources/fxml/MainView.fxml
@@ -34,7 +34,7 @@
 				prefHeight="346.0" vgrow="SOMETIMES" />
 		</rowConstraints>
 		<children>
-			<fx:include fx:id="dataContent" source="DataView.fxml"
+			<fx:include fx:id="strandContent" source="StrandView.fxml"
 				GridPane.columnIndex="0" GridPane.rowIndex="1" />
 			<fx:include fx:id="treeContent" source="TreeView.fxml"
 				GridPane.columnIndex="1" GridPane.rowIndex="1" />
